requirements.txt
django-import-export==1.1.0
djangorestframework==3.9.0
django-tables2==2.0.2
pandas==0.23.4